requirements.txt
asgiref==3.4.1
astroid==2.9.3
attrs==21.4.0
backports.zoneinfo==0.2.1
black==21.12b0
click==8.0.3
Django==4.0.1
djangorestframework==3.13.1
drf-spectacular==0.21.1
gunicorn==20.1.0
importlib-resources==5.4.0
inflection==0.5.1
isort==5.10.1
jsonschema==4.3.3
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
wrapt==1.13.3
zipp==3.7.0
